Change back to float32

+++ trainer/default_type.py
@@ -1,3 +1,3 @@
 import tensorflow as tf
 
-my_tf_float = tf.float16
+my_tf_float = tf.float32
